% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facsimile.R
\name{get_facsimile}
\alias{get_facsimile}
\alias{get_ubpr_facsimile}
\title{Retrieve Facsimile}
\usage{
get_facsimile(
  user_id = Sys.getenv("FFIEC_USER_ID"),
  bearer_token = Sys.getenv("FFIEC_BEARER_TOKEN"),
  reporting_period_end_date,
  fi_id_type = c("ID_RSSD", "FDICCertNumber", "OCCChartNumber", "OTSDockNumber"),
  fi_id
)

get_ubpr_facsimile(
  user_id = Sys.getenv("FFIEC_USER_ID"),
  bearer_token = Sys.getenv("FFIEC_BEARER_TOKEN"),
  reporting_period_end_date,
  fi_id_type = c("ID_RSSD", "FDICCertNumber", "OCCChartNumber", "OTSDockNumber"),
  fi_id
)
}
\arguments{
\item{user_id}{(String) The UserID for authenticating against the FFIEC API}

\item{bearer_token}{(String) The Bearer Token for authenticating against the
FFIEC API}

\item{reporting_period_end_date}{(String, character vector, Date, or Date
vector) One or more reporting period end dates. Character values must be
formatted as "MM/DD/YYYY". Date objects are also accepted and will be
coerced to the required format automatically.}

\item{fi_id_type}{(String) The type of identifier being provided; one of
\code{c("ID_RSSD", "FDICCertNumber", "OCCChartNumber", "OTSDockNumber")};
default is "ID_RSSD"}

\item{fi_id}{(String or character vector) One or more financial institution
identifiers (can also be supplied as an integer vector)}
}
\value{
A tibble containing the facsimile data.
}
\description{
Retrieves Call Report or UBPR facsimile data from the FFIEC
Central Data Repository API for the requested financial institution.
}
\examples{
if (!no_creds_available()) {
  # Assume you have set the following environment variables:
  # - FFIEC_USER_ID
  # - FFIEC_BEARER_TOKEN

  # Retrieve facsimile data for reporting period 2025-03-31 for institution
  # with ID RSSD "480228"
  get_facsimile(
    reporting_period_end_date = "03/31/2025",
    fi_id = 480228
  )

  # Retrieve UBPR facsimile data for reporting period 2025-03-31 for
  # instutition with FDIC Cert Number "3510"
  get_ubpr_facsimile(
    reporting_period_end_date = "03/31/2025",
    fi_id_type = "FDICCertNumber",
    fi_id = "3510"
  )

  # Retrieve facsimile data for reporting periods 2025-03-31 and 2025-06-30
  # for institutions with ID RSSD of "480228" and "451965"
  get_facsimile(
    reporting_period_end_date = c("03/31/2025", "06/30/2025"),
    fi_id = c("480228", "451965")
  )

  # Retrieve UBPR data for reporting periods 2025-03-31 and 2025-06-30
  # for institution with ID RSSD of "480228"
  get_ubpr_facsimile(
    reporting_period_end_date = c("03/31/2025", "06/30/2025"),
    fi_id = 480228
  )
}
}
\references{
\url{https://cdr.ffiec.gov/public/Files/SIS611_-_Retrieve_Public_Data_via_Web_Service.pdf}
}
