% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{no_creds_available}
\alias{no_creds_available}
\title{Handle missing UserID / Bearer Token values without throwing an error for
unit testing purposes}
\usage{
no_creds_available(
  user_id = Sys.getenv("FFIEC_USER_ID"),
  bearer_token = Sys.getenv("FFIEC_BEARER_TOKEN")
)
}
\arguments{
\item{user_id}{(String) The UserID for authenticating against the FFIEC API}

\item{bearer_token}{(String) The Bearer Token for authenticating against the
FFIEC API}
}
\value{
(Logical) `FALSE` if a valid `user_id` and `bearer_token` are
  available; otherwise `TRUE`.
}
\description{
Handle missing UserID / Bearer Token values without throwing an error for
unit testing purposes
}
\details{
Intended for internal use.
}
